# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProfilePrivacyBio = typeof ProfilePrivacyBio[keyof typeof ProfilePrivacyBio];


export const ProfilePrivacyBio = {
  public: 'public',
  friends: 'friends',
  private: 'private',
} as const;

export type ProfilePrivacyBirthday = typeof ProfilePrivacyBirthday[keyof typeof ProfilePrivacyBirthday];


export const ProfilePrivacyBirthday = {
  public: 'public',
  friends: 'friends',
  private: 'private',
} as const;

export type ProfilePrivacyLocation = typeof ProfilePrivacyLocation[keyof typeof ProfilePrivacyLocation];


export const ProfilePrivacyLocation = {
  public: 'public',
  friends: 'friends',
  private: 'private',
} as const;

export type ProfilePrivacyOccupation = typeof ProfilePrivacyOccupation[keyof typeof ProfilePrivacyOccupation];


export const ProfilePrivacyOccupation = {
  public: 'public',
  friends: 'friends',
  private: 'private',
} as const;

export type ProfilePrivacyWebsite = typeof ProfilePrivacyWebsite[keyof typeof ProfilePrivacyWebsite];


export const ProfilePrivacyWebsite = {
  public: 'public',
  friends: 'friends',
  private: 'private',
} as const;

export type ProfilePrivacyPhone = typeof ProfilePrivacyPhone[keyof typeof ProfilePrivacyPhone];


export const ProfilePrivacyPhone = {
  public: 'public',
  friends: 'friends',
  private: 'private',
} as const;

export type ProfilePrivacyEmail = typeof ProfilePrivacyEmail[keyof typeof ProfilePrivacyEmail];


export const ProfilePrivacyEmail = {
  public: 'public',
  friends: 'friends',
  private: 'private',
} as const;

export type ProfilePrivacyAbout = typeof ProfilePrivacyAbout[keyof typeof ProfilePrivacyAbout];


export const ProfilePrivacyAbout = {
  public: 'public',
  friends: 'friends',
  private: 'private',
} as const;

export type ProfilePrivacyPronouns = typeof ProfilePrivacyPronouns[keyof typeof ProfilePrivacyPronouns];


export const ProfilePrivacyPronouns = {
  public: 'public',
  friends: 'friends',
  private: 'private',
} as const;

export type ProfilePrivacySocialLinks = typeof ProfilePrivacySocialLinks[keyof typeof ProfilePrivacySocialLinks];


export const ProfilePrivacySocialLinks = {
  public: 'public',
  friends: 'friends',
  private: 'private',
} as const;

export interface ProfilePrivacy {
  bio: ProfilePrivacyBio;
  birthday: ProfilePrivacyBirthday;
  location: ProfilePrivacyLocation;
  occupation: ProfilePrivacyOccupation;
  website: ProfilePrivacyWebsite;
  phone: ProfilePrivacyPhone;
  email: ProfilePrivacyEmail;
  about: ProfilePrivacyAbout;
  pronouns: ProfilePrivacyPronouns;
  socialLinks: ProfilePrivacySocialLinks;
}

export interface Profile {
  /** User ID (owner) */
  id: string;
  name?: string;
  username?: string;
  avatarColor?: string;
  avatarUri?: string;
  pronouns?: string;
  bio?: string;
  about?: string;
  birthday?: string;
  location?: string;
  occupation?: string;
  website?: string;
  phone?: string;
  email?: string;
  socialTwitter?: string;
  socialInstagram?: string;
  socialLinkedin?: string;
  onboarded?: boolean;
  privacy: ProfilePrivacy;
  createdAt: string;
  updatedAt: string;
}

export interface UpdateProfile {
  name?: string;
  username?: string;
  avatarColor?: string;
  avatarUri?: string;
  pronouns?: string;
  bio?: string;
  about?: string;
  birthday?: string;
  location?: string;
  occupation?: string;
  website?: string;
  phone?: string;
  email?: string;
  socialTwitter?: string;
  socialInstagram?: string;
  socialLinkedin?: string;
  onboarded?: boolean;
  privacy?: ProfilePrivacy;
  /** Client timestamp for conflict detection */
  updatedAt?: string;
}

export interface Error {
  error: string;
}