# lib/api-zod/src/generated/types/updateProfile.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProfilePrivacy } from './profilePrivacy';

export interface UpdateProfile {
  name?: string;
  username?: string;
  avatarColor?: string;
  avatarUri?: string;
  pronouns?: string;
  bio?: string;
  about?: string;
  birthday?: string;
  location?: string;
  occupation?: string;
  website?: string;
  phone?: string;
  email?: string;
  socialTwitter?: string;
  socialInstagram?: string;
  socialLinkedin?: string;
  onboarded?: boolean;
  privacy?: ProfilePrivacy;
  /** Client timestamp for conflict detection */
  updatedAt?: Date;
}